Trivia Game with Corresponding CSV
Trivia Game inspired by "Who Wants To Be a Millionaire" show. Utilizes a GUI in which the player is presented with a randomized question from the Questions CSV, then given the corresponding four answers to choose from. Scoring is tracked by a monetary scoring tower, in which the value is highlighted green for every successive correct answer. If a wrong answer is selected, the user is informed of the correct answer, and the program ends. The player wins if they answer ten correct answers in a row.